backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
SQLAlchemy==2.0.29
pydantic[email]==1.10.15
python-dotenv==1.0.1
email-validator==2.1.1
pytest==8.1.1
httpx==0.27.0

# Dependencias para autenticación y seguridad
PyJWT==2.8.0
bcrypt==4.1.2
python-multipart==0.0.9

# Dependencias para PostgreSQL (producción)
psycopg2-binary==2.9.9

# Dependencias para generación de reportes
reportlab==4.1.0
openpyxl==3.1.2

# Dependencias para validación adicional
phonenumbers==8.13.32
